requirements.txt
dearpygui
discord.py
pyyaml
watchdog
Pillow
